Add hostgroup decorator and testing setup.py
This adds a @hostgroup decorator for defining hostgroups which can then
be referred to by the command line tool. It also adds a setup.py file
which sets up a `paro` cli tool, and the corresponding paro.py file
which provides the commands for said tool. Currently it manually
implements a localhost group and a hostname command which takes an
optional `--fqdn` flag.